feat: remove unnesecary lambda
