+++ stacks/singi-bot/webHandler.ts
@@ -8,7 +8,6 @@ import {
   Runtime,
 } from "aws-cdk-lib/aws-lambda"
 import { Duration } from "aws-cdk-lib"
-import { Queue } from "aws-cdk-lib/aws-sqs"
 import { RetentionDays } from "aws-cdk-lib/aws-logs"
 import { Commonparams } from "../../lib/singi-bot-stack"
 import { HttpApi, HttpMethod } from "aws-cdk-lib/aws-apigatewayv2"
@@ -24,10 +23,8 @@ export class WebHandler {
 
   readonly lamdaExtension: ParamsAndSecretsLayerVersion
   readonly messageTable: ITable
-  readonly inqueLambda: NodejsFunction
-  readonly responseLambda: NodejsFunction
+  readonly lambdaFunc: NodejsFunction
   readonly api: HttpApi
-  readonly que: Queue
 
   constructor(
     prefix: string,
@@ -38,19 +35,11 @@ export class WebHandler {
     this.construct = construct
     this.commonParams = commonparams
 
-    this.que = this.createQue()
     this.lamdaExtension = this.getLambdaExtension()
     this.messageTable = this.getDynamoDBTable()
-    this.inqueLambda = this.crearteInqueLambda()
-    this.responseLambda = this.createResponseLambda()
+    this.lambdaFunc = this.crearteLambda()
     this.api = this.createGateway()
     this.grant()
-  }
-
-  private createQue(): Queue {
-    return new Queue(this.construct, `${this.prefix}-que`, {
-      queueName: `${this.prefix}-que`,
-    })
   }
 
   private getLambdaExtension(): ParamsAndSecretsLayerVersion {
@@ -71,59 +60,31 @@ export class WebHandler {
     )
   }
 
-  private crearteInqueLambda(): NodejsFunction {
+  private crearteLambda(): NodejsFunction {
     const funcName = `${this.prefix}-inque`
-    const entry = path.join(process.cwd(), "lambda", "inqueLambda.ts")
+    const entry = path.join(process.cwd(), "lambda", "lambda.ts")
 
     return new NodejsFunction(this.construct, funcName, {
       entry,
       functionName: funcName,
       runtime: Runtime.NODEJS_20_X,
       timeout: Duration.seconds(10),
-      memorySize: 1024,
-      environment: {
-        QUE_URL: this.que.queueUrl,
-        SLACK_BOT_TOKEN_PARAM: this.commonParams.slackBotToken,
-        SLACK_SIGNING_SECRET_PARAM: this.commonParams.slackSigninSecret,
-        MESSAGE_TABLE_NAME: this.commonParams.messageTableName,
-      },
-      paramsAndSecrets: this.lamdaExtension,
-      // TODO: log保存期間と保存場所を変更する
-      logRetention: RetentionDays.ONE_DAY,
-    })
-  }
-
-  private createResponseLambda(): NodejsFunction {
-    const funcName = `${this.prefix}-response`
-    const entry = path.join(process.cwd(), "lambda", "singiLambda.ts")
-
-    const func = new NodejsFunction(this.construct, funcName, {
-      entry,
-      functionName: funcName,
-      runtime: Runtime.NODEJS_20_X,
-      timeout: Duration.seconds(10),
       memorySize: 2056,
-      // TODO: log保存期間と保存場所を変更する
-      logRetention: RetentionDays.ONE_DAY,
       environment: {
         SLACK_BOT_TOKEN_PARAM: this.commonParams.slackBotToken,
         SLACK_SIGNING_SECRET_PARAM: this.commonParams.slackSigninSecret,
         MESSAGE_TABLE_NAME: this.commonParams.messageTableName,
       },
       paramsAndSecrets: this.lamdaExtension,
+      // TODO: log保存期間と保存場所を変更する
+      logRetention: RetentionDays.ONE_DAY,
     })
-
-    func.addEventSourceMapping("SqsEventSource", {
-      eventSourceArn: this.que.queueArn,
-    })
-
-    return func
   }
 
   private createGateway(): HttpApi {
     const httpLambdaIntegRation = new HttpLambdaIntegration(
       `${this.prefix}-integ`,
-      this.inqueLambda
+      this.lambdaFunc
     )
     const api = new HttpApi(this.construct, `${this.prefix}-gateway`, {
       apiName: `${this.prefix}-gateway`,
@@ -138,16 +99,13 @@ export class WebHandler {
   }
 
   private grant() {
-    // Lambda <=> Que間の権限を付与
-    this.que.grantConsumeMessages(this.responseLambda)
-    this.que.grantSendMessages(this.inqueLambda)
     // BedrockRuntimeのinvokeModelを呼び出すための権限を付与
     const policy = new PolicyStatement({
       effect: Effect.ALLOW,
       actions: ["bedrock:InvokeModel"],
       resources: ["*"],
     })
-    this.responseLambda.addToRolePolicy(policy)
+    this.lambdaFunc.addToRolePolicy(policy)
     // Lambdaに SSM ParameterStoreへのアクセス権限を付与
     const slackSigninSecretParam =
       StringParameter.fromStringParameterAttributes(
@@ -166,12 +124,10 @@ export class WebHandler {
     )
     
 
-    slackSigninSecretParam.grantRead(this.inqueLambda)
-    slackBotTokenParam.grantRead(this.inqueLambda)
-    slackBotTokenParam.grantRead(this.responseLambda)
+    slackSigninSecretParam.grantRead(this.lambdaFunc)
+    slackBotTokenParam.grantRead(this.lambdaFunc)
     
-    this.messageTable.grantReadWriteData(this.inqueLambda)
-    this.messageTable.grantReadWriteData(this.responseLambda)
+    this.messageTable.grantReadWriteData(this.lambdaFunc)
     const indexPolicy = new PolicyStatement({
       effect: Effect.ALLOW,
       actions: ["dynamodb:Query"],
@@ -180,8 +136,6 @@ export class WebHandler {
       ],
     });
     
-    this.inqueLambda.addToRolePolicy(indexPolicy);
-    this.responseLambda.addToRolePolicy(indexPolicy);
-    this.responseLambda.addToRolePolicy(indexPolicy)
+    this.lambdaFunc.addToRolePolicy(indexPolicy);
   }
 }
